♻️(state): state is now modified with methods on `State`, and `statics.rs` is now gone
